Nexus-Agentic/requirements.txt
crewai==0.11.0
python-dotenv==1.0.0
openai>=1.7.1,<2.0.0
duckduckgo-search==3.9.9
langchain>=0.1.0,<0.2.0
langchain-openai>=0.0.2.post1
langchain-community>=0.0.38